# KNOX-2134 - Checking if password is available via local alias service before going to fetch it from remote ZK server