--- a/gateway-server/src/test/java/org/apache/knox/gateway/services/security/impl/ZookeeperRemoteAliasMonitorTest.java
+++ b/gateway-server/src/test/java/org/apache/knox/gateway/services/security/impl/ZookeeperRemoteAliasMonitorTest.java
@@ -189,7 +189,7 @@ public class ZookeeperRemoteAliasMonitorTest {
         .andReturn(new ArrayList<>()).anyTimes();
 
     EasyMock.expect(defaultAlias.getPasswordFromAliasForCluster(expectedClusterName, preferRemoteAlias))
-        .andReturn("thisiswrong".toCharArray()).anyTimes();
+        .andReturn(preferRemoteAliasClearPassword.toCharArray()).anyTimes();
 
     EasyMock.replay(defaultAlias);