## What changes were proposed in this pull request?

Formatted the code and did a minor cleanup to make it easier to read and understand the code.
Apart from code cleanup, I changed the way how passwords are fetched in `ZookeeperRemoteAliasService`: before going to the remote ZK server we first check if the desired password is available through the local alias service. If so; we skip the remote check. This came with a separate commit to make it easier for review (see [here](https://github.com/apache/knox/pull/218/commits/2703684c0d4900340ac18741a731869c4199b451)).

## How was this patch tested?

Added a JUnit test to measure performance (ignoring it from automatic execution, but can be executed manually by commenting out the `Ignore` annotation): added and fetched 200 aliases/passwords with and without my changes in place with the following results

|Without cache (ms) |  With Cache (ms) |
|--|--|
| 30841<br>31335<br>32762<br>31646<br>31736<br>31641<br>31204<br>34219<br>33720<br>33297 |  16715<br>16265<br>17469<br>17312<br>17482<br>20990<br>17343<br>19637<br>19240<br>17705 |