@smolnar82 2019-12-16
@pzampino Please review my changes when you have time. Thanks!